+++ libc/bsearch.h
@@ -14,4 +14,4 @@
 #include "../cpu/type.h"
 #include "null.h"
 
-void* bsearch(const void* key, const void* base, u16 num, u16 size, int(*compare)(const void* key, const void* elt));
+void* bsearch(const void* key, const void* base, size_t num, size_t size, int(*compare)(const void* key, const void* elt));
